+++ src/adc.rs
@@ -29,10 +29,7 @@ pub fn prepare<U: Unsigned, P: Pin<U>>(pin: &mut P) {
     // GPIO, and we need to disable the keeper to prevent signal
     // jumps.
     super::alternate(pin, <P as super::gpio::Pin>::ALT);
-    super::configure(
-        pin,
-        super::Config::modify().set_pull_keep(super::PullKeep::Disabled),
-    );
+    super::configure(pin, super::Config::modify().set_pull_keeper(None));
 }
 
 #[allow(unused)] // Used in chip-specific modules...