+++ src/lib.rs
@@ -101,9 +101,10 @@ pub mod sai;
 use core::ptr;
 
 pub use config::{
-    configure, Config, DriveStrength, Hysteresis, OpenDrain, PullKeep, PullKeepSelect, PullUpDown,
-    SlewRate, Speed,
+    configure, Config, DriveStrength, Hysteresis, OpenDrain, PullKeeper, SlewRate, Speed,
 };
+
+pub use config::{PullKeep, PullKeepSelect, PullUpDown};
 
 /// Re-export of top-level components, without the chip-specific modules.
 ///
@@ -122,9 +123,11 @@ pub use config::{
 /// ```
 pub mod prelude {
     pub use crate::config::{
-        configure, Config, DriveStrength, Hysteresis, OpenDrain, PullKeep, PullKeepSelect,
-        PullUpDown, SlewRate, Speed,
+        configure, Config, DriveStrength, Hysteresis, OpenDrain, PullKeeper, SlewRate, Speed,
     };
+
+    pub use crate::config::{PullKeep, PullKeepSelect, PullUpDown};
+
     pub use crate::{
         consts, flexpwm, gpio, lpi2c, lpspi, lpuart, Daisy, ErasedPad, Pad, WrongPadError,
     };